fix: restore CI environment compatibility
